Add files via upload
These are Matlab codes which were used/will be used to compare to the c++ code.

Hypergeom_Check.m calculates the hypergeometric function to compare to the first cpp file. 

errorCheck.m is a prototyped mixed fixed-point Newton's method numerical method which is used to calculate gammas across a slit. This still needs to be converted to a cpp file.